// FrontEnd/src/app/services/api-backend/model/producto.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Lote } from './lote';
import { MedicamentosRecetum } from './medicamentosRecetum';
import { MedicamentosCaso } from './medicamentosCaso';
import { ProductosFactura } from './productosFactura';


export interface Producto { 
    idProducto?: number;
    nombre?: string | null;
    imagen?: string | null;
    nombrearchivo?: string | null;
    archivobase64?: string | null;
    lotes?: Array<Lote> | null;
    medicamentosCasos?: Array<MedicamentosCaso> | null;
    medicamentosReceta?: Array<MedicamentosRecetum> | null;
    productosFacturas?: Array<ProductosFactura> | null;
}